Misc.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@44326 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/isc_sync.cpp
@@ -3951,7 +3951,7 @@ static bool make_object_name(TEXT* buffer, size_t bufsize,
  **************************************/
 	char hostname[64];
 	const int rc = snprintf(buffer, bufsize, object_name, ISC_get_host(hostname, sizeof(hostname)));
-	if (rc == bufsize || rc <= 0)
+	if (size_t(rc) == bufsize || rc <= 0)
 	{
 		SetLastError(ERROR_FILENAME_EXCED_RANGE);
 		return false;
